requirements.txt
celery==5.0.0
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.6
redis==3.5.3
